move reducer and state to app
