+++ src/labelsContainer.js
@@ -1,45 +1,25 @@
 import { useReducer } from "react"
 import Label from "./label"
-const initalState = { Session: 1500, Break: 300}
-
-const reducer = (state, action) => {
-    let tempState = { ...state }
-    const minute = 60
-    const maxCount = 3600
-    const minCount = 0
-    switch (action.type) {
-        case 'increment':
-            tempState[action.property] =  state[action.property] + minute 
-            return tempState[action.property] > maxCount ? state : tempState
-        case 'decrement':
-            tempState[action.property] = state[action.property] - minute 
-            return tempState[action.property] < minCount ? state : tempState
-        case 'reset':
-            return initalState
-    }
-
-}
 
 
-const LabelsContainer = () => {
-    const [state, dispatch] = useReducer(reducer, initalState)
-    const handlePlusMinus = (actionType, actionProperty) => {
 
-        dispatch({ type: actionType, property: actionProperty})
-    }
+
+const LabelsContainer = (props) => {
+    const { state, handlePlusMinus } = props
+
     return (
         <div id="labels-part">
             <Label
-            id={'Session'}
-            state={state}
-            handlePlusMinus={handlePlusMinus} 
+                id={'Session'}
+                state={state}
+                handlePlusMinus={handlePlusMinus}
             />
             <Label
-            id={'Break'}
-            state={state}
-            handlePlusMinus={handlePlusMinus} 
+                id={'Break'}
+                state={state}
+                handlePlusMinus={handlePlusMinus}
             />
-            
+
         </div>
     )
 }
